+++ include/Mesh.h
@@ -62,6 +62,7 @@ public:
     // render the mesh
     void Draw(Shader &shader)
     {
+        //shader.use();
         // bind appropriate textures
         unsigned int diffuseNr  = 1;
         unsigned int specularNr = 1;